[DataAvailability] Refactor pipeline and core to make testing easier
