+++ module/executiondatasync/optimistic_sync/pipeline/core_impl_test.go
@@ -19,8 +19,8 @@ import (
 	"github.com/onflow/flow-go/utils/unittest/fixtures"
 )
 
-// CoreImplSuite is a test suite for testing the CoreImpl.
-type CoreImplSuite struct {
+// CoreSuite is a test suite for testing the Core.
+type CoreSuite struct {
 	suite.Suite
 	execDataRequester             *reqestermock.ExecutionDataRequester
 	txResultErrMsgsRequester      *txerrmsgsmock.Requester
@@ -35,12 +35,12 @@ type CoreImplSuite struct {
 	latestPersistedSealedResult   *storagemock.LatestPersistedSealedResult
 }
 
-func TestCoreImplSuiteSuite(t *testing.T) {
+func TestCoreSuiteSuite(t *testing.T) {
 	t.Parallel()
-	suite.Run(t, new(CoreImplSuite))
+	suite.Run(t, new(CoreSuite))
 }
 
-func (c *CoreImplSuite) SetupTest() {
+func (c *CoreSuite) SetupTest() {
 	t := c.T()
 
 	c.execDataRequester = reqestermock.NewExecutionDataRequester(t)
@@ -48,11 +48,11 @@ func (c *CoreImplSuite) SetupTest() {
 	c.txResultErrMsgsRequestTimeout = 100 * time.Millisecond
 }
 
-// createTestCoreImpl creates a CoreImpl instance with mocked dependencies for testing.
+// createTestCore creates a Core instance with mocked dependencies for testing.
 //
-// Returns a configured CoreImpl ready for testing.
-func (c *CoreImplSuite) createTestCoreImpl(tf *testFixture) *CoreImpl {
-	core, err := NewCoreImpl(
+// Returns a configured Core ready for testing.
+func (c *CoreSuite) createTestCore(tf *testFixture) *Core {
+	core, err := NewCore(
 		unittest.Logger(),
 		tf.exeResult,
 		tf.block,
@@ -128,12 +128,12 @@ func generateFixture(g *fixtures.GeneratorSuite) *testFixture {
 	}
 }
 
-func (c *CoreImplSuite) TestCoreImpl_Constructor() {
+func (c *CoreSuite) TestCore_Constructor() {
 	block := unittest.BlockFixture()
 	executionResult := unittest.ExecutionResultFixture(unittest.WithBlock(block))
 
 	c.Run("happy path", func() {
-		core, err := NewCoreImpl(
+		core, err := NewCore(
 			unittest.Logger(),
 			executionResult,
 			block,
@@ -154,7 +154,7 @@ func (c *CoreImplSuite) TestCoreImpl_Constructor() {
 	})
 
 	c.Run("block ID mismatch", func() {
-		core, err := NewCoreImpl(
+		core, err := NewCore(
 			unittest.Logger(),
 			executionResult,
 			unittest.BlockFixture(),
@@ -175,9 +175,9 @@ func (c *CoreImplSuite) TestCoreImpl_Constructor() {
 	})
 }
 
-// TestCoreImpl_Download tests the Download method retrieves execution data and transaction error
+// TestCore_Download tests the Download method retrieves execution data and transaction error
 // messages.
-func (c *CoreImplSuite) TestCoreImpl_Download() {
+func (c *CoreSuite) TestCore_Download() {
 	ctx := context.Background()
 	g := fixtures.NewGeneratorSuite()
 
@@ -191,7 +191,7 @@ func (c *CoreImplSuite) TestCoreImpl_Download() {
 
 	c.Run("successful download", func() {
 		tf := generateFixture(g)
-		core := c.createTestCoreImpl(tf)
+		core := c.createTestCore(tf)
 
 		c.execDataRequester.On("RequestExecutionData", mock.Anything).Return(tf.execData, nil).Once()
 		c.txResultErrMsgsRequester.On("Request", mock.Anything).Return(tf.txErrMsgs, nil).Once()
@@ -209,7 +209,7 @@ func (c *CoreImplSuite) TestCoreImpl_Download() {
 
 	c.Run("execution data request error", func() {
 		tf := generateFixture(g)
-		core := c.createTestCoreImpl(tf)
+		core := c.createTestCore(tf)
 
 		expectedErr := fmt.Errorf("test execution data request error")
 
@@ -229,7 +229,7 @@ func (c *CoreImplSuite) TestCoreImpl_Download() {
 
 	c.Run("transaction result error messages request error", func() {
 		tf := generateFixture(g)
-		core := c.createTestCoreImpl(tf)
+		core := c.createTestCore(tf)
 
 		expectedErr := fmt.Errorf("test tx error messages request error")
 
@@ -249,7 +249,7 @@ func (c *CoreImplSuite) TestCoreImpl_Download() {
 
 	c.Run("context cancellation", func() {
 		tf := generateFixture(g)
-		core := c.createTestCoreImpl(tf)
+		core := c.createTestCore(tf)
 
 		ctx, cancel := context.WithCancel(context.Background())
 		cancel()
@@ -275,7 +275,7 @@ func (c *CoreImplSuite) TestCoreImpl_Download() {
 
 	c.Run("txResultErrMsgsRequestTimeout expiration", func() {
 		tf := generateFixture(g)
-		core := c.createTestCoreImpl(tf)
+		core := c.createTestCore(tf)
 
 		c.execDataRequester.On("RequestExecutionData", mock.Anything).Return(tf.execData, nil).Once()
 
@@ -302,7 +302,7 @@ func (c *CoreImplSuite) TestCoreImpl_Download() {
 
 	c.Run("Download after Abandon returns an error", func() {
 		tf := generateFixture(g)
-		core := c.createTestCoreImpl(tf)
+		core := c.createTestCore(tf)
 
 		core.Abandon()
 		c.Nil(core.workingData)
@@ -312,8 +312,8 @@ func (c *CoreImplSuite) TestCoreImpl_Download() {
 	})
 }
 
-// TestCoreImpl_Index tests the Index method which processes downloaded data.
-func (c *CoreImplSuite) TestCoreImpl_Index() {
+// TestCore_Index tests the Index method which processes downloaded data.
+func (c *CoreSuite) TestCore_Index() {
 	ctx := context.Background()
 	g := fixtures.NewGeneratorSuite()
 
@@ -322,7 +322,7 @@ func (c *CoreImplSuite) TestCoreImpl_Index() {
 	c.txResultErrMsgsRequester.On("Request", mock.Anything).Return(tf.txErrMsgs, nil)
 
 	c.Run("successful indexing", func() {
-		core := c.createTestCoreImpl(tf)
+		core := c.createTestCore(tf)
 
 		err := core.Download(ctx)
 		c.Require().NoError(err)
@@ -345,7 +345,7 @@ func (c *CoreImplSuite) TestCoreImpl_Index() {
 	})
 
 	c.Run("indexer constructor error", func() {
-		core := c.createTestCoreImpl(tf)
+		core := c.createTestCore(tf)
 		core.block = g.Blocks().Fixture()
 
 		err := core.Download(ctx)
@@ -357,7 +357,7 @@ func (c *CoreImplSuite) TestCoreImpl_Index() {
 	})
 
 	c.Run("failed to index block", func() {
-		core := c.createTestCoreImpl(tf)
+		core := c.createTestCore(tf)
 
 		err := core.Download(ctx)
 		c.Require().NoError(err)
@@ -370,7 +370,7 @@ func (c *CoreImplSuite) TestCoreImpl_Index() {
 	})
 
 	c.Run("failed to validate transaction result error messages", func() {
-		core := c.createTestCoreImpl(tf)
+		core := c.createTestCore(tf)
 
 		err := core.Download(ctx)
 		c.Require().NoError(err)
@@ -386,7 +386,7 @@ func (c *CoreImplSuite) TestCoreImpl_Index() {
 	})
 
 	c.Run("Index after Abandon returns an error", func() {
-		core := c.createTestCoreImpl(tf)
+		core := c.createTestCore(tf)
 
 		core.Abandon()
 		c.Nil(core.workingData)
@@ -396,7 +396,7 @@ func (c *CoreImplSuite) TestCoreImpl_Index() {
 	})
 
 	c.Run("Index before Download returns an error", func() {
-		core := c.createTestCoreImpl(tf)
+		core := c.createTestCore(tf)
 
 		err := core.Index()
 		c.ErrorContains(err, "downloading is not complete")
@@ -404,8 +404,8 @@ func (c *CoreImplSuite) TestCoreImpl_Index() {
 	})
 }
 
-// TestCoreImpl_Persist tests the Persist method which persists indexed data to storages and database.
-func (c *CoreImplSuite) TestCoreImpl_Persist() {
+// TestCore_Persist tests the Persist method which persists indexed data to storages and database.
+func (c *CoreSuite) TestCore_Persist() {
 	t := c.T()
 	ctx := context.Background()
 	g := fixtures.NewGeneratorSuite()
@@ -428,7 +428,7 @@ func (c *CoreImplSuite) TestCoreImpl_Persist() {
 
 	c.Run("successful persistence of empty data", func() {
 		resetMocks()
-		core := c.createTestCoreImpl(tf)
+		core := c.createTestCore(tf)
 
 		err := core.Download(ctx)
 		c.Require().NoError(err)
@@ -463,7 +463,7 @@ func (c *CoreImplSuite) TestCoreImpl_Persist() {
 		expectedErr := fmt.Errorf("test persisting registers failure")
 
 		resetMocks()
-		core := c.createTestCoreImpl(tf)
+		core := c.createTestCore(tf)
 
 		err := core.Download(ctx)
 		c.Require().NoError(err)
@@ -482,7 +482,7 @@ func (c *CoreImplSuite) TestCoreImpl_Persist() {
 		expectedErr := fmt.Errorf("test persisting events failure")
 
 		resetMocks()
-		core := c.createTestCoreImpl(tf)
+		core := c.createTestCore(tf)
 
 		err := core.Download(ctx)
 		c.Require().NoError(err)
@@ -507,7 +507,7 @@ func (c *CoreImplSuite) TestCoreImpl_Persist() {
 
 	c.Run("Persist after Abandon returns an error", func() {
 		resetMocks()
-		core := c.createTestCoreImpl(tf)
+		core := c.createTestCore(tf)
 
 		core.Abandon()
 		c.Nil(core.workingData)
@@ -518,7 +518,7 @@ func (c *CoreImplSuite) TestCoreImpl_Persist() {
 
 	c.Run("Persist before Index returns an error", func() {
 		resetMocks()
-		core := c.createTestCoreImpl(tf)
+		core := c.createTestCore(tf)
 		err := core.Persist()
 		c.ErrorContains(err, "indexing is not complete")
 	})
