+++ module/executiondatasync/optimistic_sync/pipeline/pipeline_functional_test.go
@@ -18,6 +18,7 @@ import (
 	"github.com/onflow/flow-go/model/flow"
 	"github.com/onflow/flow-go/module"
 	"github.com/onflow/flow-go/module/executiondatasync/execution_data"
+	"github.com/onflow/flow-go/module/executiondatasync/optimistic_sync"
 	"github.com/onflow/flow-go/module/metrics"
 	reqestermock "github.com/onflow/flow-go/module/state_synchronization/requester/mock"
 	"github.com/onflow/flow-go/storage"
@@ -52,7 +53,7 @@ type PipelineFunctionalSuite struct {
 	headers                       *store.Headers
 	results                       *store.ExecutionResults
 	persistentLatestSealedResult  *store.LatestPersistedSealedResult
-	core                          *CoreImpl
+	core                          *Core
 	block                         *flow.Block
 	executionResult               *flow.ExecutionResult
 	metrics                       module.CacheMetrics
@@ -168,7 +169,7 @@ func (p *PipelineFunctionalSuite) SetupTest() {
 	p.txResultErrMsgsRequestTimeout = DefaultTxResultErrMsgsRequestTimeout
 
 	p.config = PipelineConfig{
-		parentState: StateWaitingPersist,
+		parentState: optimistic_sync.StateWaitingPersist,
 	}
 
 	//  generate expected data based on the fixtures
@@ -230,14 +231,14 @@ func (p *PipelineFunctionalSuite) TestPipelineCompletesSuccessfully() {
 	p.execDataRequester.On("RequestExecutionData", mock.Anything).Return(p.expectedExecutionData, nil).Once()
 	p.txResultErrMsgsRequester.On("Request", mock.Anything).Return(p.expectedTxResultErrMsgs, nil).Once()
 
-	p.WithRunningPipeline(func(pipeline Pipeline, updateChan chan State, errChan chan error, cancel context.CancelFunc) {
-		pipeline.OnParentStateUpdated(StateComplete)
+	p.WithRunningPipeline(func(pipeline optimistic_sync.Pipeline, updateChan chan optimistic_sync.State, errChan chan error, cancel context.CancelFunc) {
+		pipeline.OnParentStateUpdated(optimistic_sync.StateComplete)
 
-		waitForStateUpdates(p.T(), updateChan, errChan, StateProcessing, StateWaitingPersist)
+		waitForStateUpdates(p.T(), updateChan, errChan, optimistic_sync.StateProcessing, optimistic_sync.StateWaitingPersist)
 
 		pipeline.SetSealed()
 
-		waitForStateUpdates(p.T(), updateChan, errChan, StateComplete)
+		waitForStateUpdates(p.T(), updateChan, errChan, optimistic_sync.StateComplete)
 
 		actualEvents, err := p.persistentEvents.ByBlockID(p.block.ID())
 		p.Require().NoError(err)
@@ -286,11 +287,11 @@ func (p *PipelineFunctionalSuite) TestPipelineDownloadError() {
 		p.execDataRequester.On("RequestExecutionData", mock.Anything).Return(nil, expectedErr).Once()
 		p.txResultErrMsgsRequester.On("Request", mock.Anything).Return(p.expectedTxResultErrMsgs, nil).Once()
 
-		p.WithRunningPipeline(func(pipeline Pipeline, updateChan chan State, errChan chan error, cancel context.CancelFunc) {
-			pipeline.OnParentStateUpdated(StateComplete)
+		p.WithRunningPipeline(func(pipeline optimistic_sync.Pipeline, updateChan chan optimistic_sync.State, errChan chan error, cancel context.CancelFunc) {
+			pipeline.OnParentStateUpdated(optimistic_sync.StateComplete)
 
 			waitForError(p.T(), errChan, expectedErr)
-			p.Assert().Equal(StateProcessing, pipeline.GetState())
+			p.Assert().Equal(optimistic_sync.StateProcessing, pipeline.GetState())
 		}, p.config)
 	})
 
@@ -300,11 +301,11 @@ func (p *PipelineFunctionalSuite) TestPipelineDownloadError() {
 		p.execDataRequester.On("RequestExecutionData", mock.Anything).Return(p.expectedExecutionData, nil).Once()
 		p.txResultErrMsgsRequester.On("Request", mock.Anything).Return(nil, expectedErr).Once()
 
-		p.WithRunningPipeline(func(pipeline Pipeline, updateChan chan State, errChan chan error, cancel context.CancelFunc) {
-			pipeline.OnParentStateUpdated(StateComplete)
+		p.WithRunningPipeline(func(pipeline optimistic_sync.Pipeline, updateChan chan optimistic_sync.State, errChan chan error, cancel context.CancelFunc) {
+			pipeline.OnParentStateUpdated(optimistic_sync.StateComplete)
 
 			waitForError(p.T(), errChan, expectedErr)
-			p.Assert().Equal(StateProcessing, pipeline.GetState())
+			p.Assert().Equal(optimistic_sync.StateProcessing, pipeline.GetState())
 		}, p.config)
 	})
 }
@@ -326,14 +327,14 @@ func (p *PipelineFunctionalSuite) TestPipelineIndexingError() {
 	expectedIndexingError := fmt.Errorf("could not perform indexing: failed to index execution data: unexpected block execution data: expected block_id=%s, actual block_id=%s",
 		p.block.ID().String(), invalidBlockID.String())
 
-	p.WithRunningPipeline(func(pipeline Pipeline, updateChan chan State, errChan chan error, cancel context.CancelFunc) {
-		pipeline.OnParentStateUpdated(StateComplete)
+	p.WithRunningPipeline(func(pipeline optimistic_sync.Pipeline, updateChan chan optimistic_sync.State, errChan chan error, cancel context.CancelFunc) {
+		pipeline.OnParentStateUpdated(optimistic_sync.StateComplete)
 
 		waitForErrorWithCustomCheckers(p.T(), errChan, func(err error) {
 			p.Require().Error(err)
 			p.Assert().Equal(expectedIndexingError.Error(), err.Error())
 		})
-		p.Assert().Equal(StateProcessing, pipeline.GetState())
+		p.Assert().Equal(optimistic_sync.StateProcessing, pipeline.GetState())
 	}, p.config)
 }
 
@@ -349,15 +350,15 @@ func (p *PipelineFunctionalSuite) TestPipelinePersistingError() {
 	p.execDataRequester.On("RequestExecutionData", mock.Anything).Return(p.expectedExecutionData, nil).Once()
 	p.txResultErrMsgsRequester.On("Request", mock.Anything).Return(p.expectedTxResultErrMsgs, nil).Once()
 
-	p.WithRunningPipeline(func(pipeline Pipeline, updateChan chan State, errChan chan error, cancel context.CancelFunc) {
-		pipeline.OnParentStateUpdated(StateComplete)
+	p.WithRunningPipeline(func(pipeline optimistic_sync.Pipeline, updateChan chan optimistic_sync.State, errChan chan error, cancel context.CancelFunc) {
+		pipeline.OnParentStateUpdated(optimistic_sync.StateComplete)
 
-		waitForStateUpdates(p.T(), updateChan, errChan, StateProcessing, StateWaitingPersist)
+		waitForStateUpdates(p.T(), updateChan, errChan, optimistic_sync.StateProcessing, optimistic_sync.StateWaitingPersist)
 
 		pipeline.SetSealed()
 
 		waitForError(p.T(), errChan, expectedError)
-		p.Assert().Equal(StateWaitingPersist, pipeline.GetState())
+		p.Assert().Equal(optimistic_sync.StateWaitingPersist, pipeline.GetState())
 	}, p.config)
 }
 
@@ -365,7 +366,7 @@ func (p *PipelineFunctionalSuite) TestPipelinePersistingError() {
 // request of execution data. It ensures that cancellation is handled properly when triggered
 // while execution data is being downloaded.
 func (p *PipelineFunctionalSuite) TestMainCtxCancellationDuringRequestingExecutionData() {
-	p.WithRunningPipeline(func(pipeline Pipeline, updateChan chan State, errChan chan error, cancel context.CancelFunc) {
+	p.WithRunningPipeline(func(pipeline optimistic_sync.Pipeline, updateChan chan optimistic_sync.State, errChan chan error, cancel context.CancelFunc) {
 		p.execDataRequester.
 			On("RequestExecutionData", mock.Anything).
 			Return(func(ctx context.Context) (*execution_data.BlockExecutionData, error) {
@@ -378,13 +379,13 @@ func (p *PipelineFunctionalSuite) TestMainCtxCancellationDuringRequestingExecuti
 		// This call marked as `Maybe()` because it may not be called depending on timing.
 		p.txResultErrMsgsRequester.On("Request", mock.Anything).Return([]flow.TransactionResultErrorMessage{}, nil).Maybe()
 
-		pipeline.OnParentStateUpdated(StateComplete)
+		pipeline.OnParentStateUpdated(optimistic_sync.StateComplete)
 
-		waitForStateUpdates(p.T(), updateChan, errChan, StateProcessing)
+		waitForStateUpdates(p.T(), updateChan, errChan, optimistic_sync.StateProcessing)
 		waitForError(p.T(), errChan, context.Canceled)
 
-		p.Assert().Equal(StateProcessing, pipeline.GetState())
-	}, PipelineConfig{parentState: StatePending})
+		p.Assert().Equal(optimistic_sync.StateProcessing, pipeline.GetState())
+	}, PipelineConfig{parentState: optimistic_sync.StatePending})
 }
 
 // TestMainCtxCancellationDuringRequestingTxResultErrMsgs tests context cancellation during
@@ -392,7 +393,7 @@ func (p *PipelineFunctionalSuite) TestMainCtxCancellationDuringRequestingExecuti
 // is cancelled during this phase, the pipeline handles the cancellation gracefully
 // and transitions to the correct state.
 func (p *PipelineFunctionalSuite) TestMainCtxCancellationDuringRequestingTxResultErrMsgs() {
-	p.WithRunningPipeline(func(pipeline Pipeline, updateChan chan State, errChan chan error, cancel context.CancelFunc) {
+	p.WithRunningPipeline(func(pipeline optimistic_sync.Pipeline, updateChan chan optimistic_sync.State, errChan chan error, cancel context.CancelFunc) {
 		// This call marked as `Maybe()` because it may not be called depending on timing.
 		p.execDataRequester.On("RequestExecutionData", mock.Anything).Return(nil, nil).Maybe()
 
@@ -405,24 +406,24 @@ func (p *PipelineFunctionalSuite) TestMainCtxCancellationDuringRequestingTxResul
 			}).
 			Once()
 
-		pipeline.OnParentStateUpdated(StateComplete)
+		pipeline.OnParentStateUpdated(optimistic_sync.StateComplete)
 
-		waitForStateUpdates(p.T(), updateChan, errChan, StateProcessing)
+		waitForStateUpdates(p.T(), updateChan, errChan, optimistic_sync.StateProcessing)
 		waitForError(p.T(), errChan, context.Canceled)
 
-		p.Assert().Equal(StateProcessing, pipeline.GetState())
-	}, PipelineConfig{parentState: StatePending})
+		p.Assert().Equal(optimistic_sync.StateProcessing, pipeline.GetState())
+	}, PipelineConfig{parentState: optimistic_sync.StatePending})
 }
 
-// TestMainCtxCancellationDuringWaitingPersist tests the pipeline's behavior when the main context is canceled during StateWaitingPersist.
+// TestMainCtxCancellationDuringWaitingPersist tests the pipeline's behavior when the main context is canceled during optimistic_sync.StateWaitingPersist.
 func (p *PipelineFunctionalSuite) TestMainCtxCancellationDuringWaitingPersist() {
 	p.execDataRequester.On("RequestExecutionData", mock.Anything).Return(p.expectedExecutionData, nil).Once()
 	p.txResultErrMsgsRequester.On("Request", mock.Anything).Return(p.expectedTxResultErrMsgs, nil).Once()
 
-	p.WithRunningPipeline(func(pipeline Pipeline, updateChan chan State, errChan chan error, cancel context.CancelFunc) {
-		pipeline.OnParentStateUpdated(StateComplete)
+	p.WithRunningPipeline(func(pipeline optimistic_sync.Pipeline, updateChan chan optimistic_sync.State, errChan chan error, cancel context.CancelFunc) {
+		pipeline.OnParentStateUpdated(optimistic_sync.StateComplete)
 
-		waitForStateUpdates(p.T(), updateChan, errChan, StateProcessing, StateWaitingPersist)
+		waitForStateUpdates(p.T(), updateChan, errChan, optimistic_sync.StateProcessing, optimistic_sync.StateWaitingPersist)
 
 		cancel()
 
@@ -430,7 +431,7 @@ func (p *PipelineFunctionalSuite) TestMainCtxCancellationDuringWaitingPersist() 
 
 		waitForError(p.T(), errChan, context.Canceled)
 
-		p.Assert().Equal(StateWaitingPersist, pipeline.GetState())
+		p.Assert().Equal(optimistic_sync.StateWaitingPersist, pipeline.GetState())
 	}, p.config)
 }
 
@@ -439,32 +440,32 @@ func (p *PipelineFunctionalSuite) TestPipelineShutdownOnParentAbandon() {
 	tests := []struct {
 		name        string
 		config      PipelineConfig
-		customSetup func(pipeline Pipeline, updateChan chan State, errChan chan error)
+		customSetup func(pipeline optimistic_sync.Pipeline, updateChan chan optimistic_sync.State, errChan chan error)
 	}{
 		{
-			name: "from StatePending",
+			name: "from optimistic_sync.StatePending",
 			config: PipelineConfig{
-				beforePipelineRun: func(pipeline *PipelineImpl) {
-					pipeline.OnParentStateUpdated(StateAbandoned)
+				beforePipelineRun: func(pipeline *Pipeline) {
+					pipeline.OnParentStateUpdated(optimistic_sync.StateAbandoned)
 				},
-				parentState: StateAbandoned,
+				parentState: optimistic_sync.StateAbandoned,
 			},
 		},
 		{
-			name: "from StateProcessing",
-			customSetup: func(pipeline Pipeline, updateChan chan State, errChan chan error) {
-				waitForStateUpdates(p.T(), updateChan, errChan, StateProcessing)
+			name: "from optimistic_sync.StateProcessing",
+			customSetup: func(pipeline optimistic_sync.Pipeline, updateChan chan optimistic_sync.State, errChan chan error) {
+				waitForStateUpdates(p.T(), updateChan, errChan, optimistic_sync.StateProcessing)
 
-				pipeline.OnParentStateUpdated(StateAbandoned)
+				pipeline.OnParentStateUpdated(optimistic_sync.StateAbandoned)
 			},
 			config: p.config,
 		},
 		{
-			name: "from StateWaitingPersist",
-			customSetup: func(pipeline Pipeline, updateChan chan State, errChan chan error) {
-				waitForStateUpdates(p.T(), updateChan, errChan, StateProcessing, StateWaitingPersist)
+			name: "from optimistic_sync.StateWaitingPersist",
+			customSetup: func(pipeline optimistic_sync.Pipeline, updateChan chan optimistic_sync.State, errChan chan error) {
+				waitForStateUpdates(p.T(), updateChan, errChan, optimistic_sync.StateProcessing, optimistic_sync.StateWaitingPersist)
 
-				pipeline.OnParentStateUpdated(StateAbandoned)
+				pipeline.OnParentStateUpdated(optimistic_sync.StateAbandoned)
 			},
 			config: p.config,
 		},
@@ -472,7 +473,7 @@ func (p *PipelineFunctionalSuite) TestPipelineShutdownOnParentAbandon() {
 
 	for _, test := range tests {
 		p.T().Run(test.name, func(t *testing.T) {
-			p.WithRunningPipeline(func(pipeline Pipeline, updateChan chan State, errChan chan error, cancel context.CancelFunc) {
+			p.WithRunningPipeline(func(pipeline optimistic_sync.Pipeline, updateChan chan optimistic_sync.State, errChan chan error, cancel context.CancelFunc) {
 				p.execDataRequester.On("RequestExecutionData", mock.Anything).Return(p.expectedExecutionData, nil).Maybe()
 				p.txResultErrMsgsRequester.On("Request", mock.Anything).Return(p.expectedTxResultErrMsgs, nil).Maybe()
 
@@ -480,10 +481,10 @@ func (p *PipelineFunctionalSuite) TestPipelineShutdownOnParentAbandon() {
 					test.customSetup(pipeline, updateChan, errChan)
 				}
 
-				waitForStateUpdates(p.T(), updateChan, errChan, StateAbandoned)
+				waitForStateUpdates(p.T(), updateChan, errChan, optimistic_sync.StateAbandoned)
 				waitForError(p.T(), errChan, nil)
 
-				p.Assert().Equal(StateAbandoned, pipeline.GetState())
+				p.Assert().Equal(optimistic_sync.StateAbandoned, pipeline.GetState())
 				p.Assert().Nil(p.core.workingData)
 			}, test.config)
 		})
@@ -491,20 +492,20 @@ func (p *PipelineFunctionalSuite) TestPipelineShutdownOnParentAbandon() {
 }
 
 type PipelineConfig struct {
-	beforePipelineRun func(pipeline *PipelineImpl)
-	parentState       State
+	beforePipelineRun func(pipeline *Pipeline)
+	parentState       optimistic_sync.State
 }
 
 // WithRunningPipeline is a test helper that initializes and starts a pipeline instance.
 // It manages the context and channels needed to run the pipeline and invokes the testFunc
 // with access to the pipeline, update channel, error channel, and cancel function.
 func (p *PipelineFunctionalSuite) WithRunningPipeline(
-	testFunc func(pipeline Pipeline, updateChan chan State, errChan chan error, cancel context.CancelFunc),
+	testFunc func(pipeline optimistic_sync.Pipeline, updateChan chan optimistic_sync.State, errChan chan error, cancel context.CancelFunc),
 	pipelineConfig PipelineConfig,
 ) {
 	var err error
 
-	p.core, err = NewCoreImpl(
+	p.core, err = NewCore(
 		p.logger,
 		p.executionResult,
 		p.block,
