+++ module/executiondatasync/optimistic_sync/pipeline/pipeline_test_utils.go
@@ -8,6 +8,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	"github.com/onflow/flow-go/module/executiondatasync/optimistic_sync"
 	osmock "github.com/onflow/flow-go/module/executiondatasync/optimistic_sync/mock"
 	"github.com/onflow/flow-go/utils/unittest"
 )
@@ -15,52 +16,52 @@ import (
 // mockStateProvider is a mock implementation of a parent state provider.
 // It tracks the current state and notifies the pipeline when the state changes.
 type mockStateProvider struct {
-	state State
+	state optimistic_sync.State
 }
 
-var _ PipelineStateProvider = (*mockStateProvider)(nil)
+var _ optimistic_sync.PipelineStateProvider = (*mockStateProvider)(nil)
 
-// NewMockStateProvider initializes a mockStateProvider with the default state StatePending.
+// NewMockStateProvider initializes a mockStateProvider with the default state optimistic_sync..
 func NewMockStateProvider() *mockStateProvider {
 	return &mockStateProvider{
-		state: StatePending,
+		state: optimistic_sync.StatePending,
 	}
 }
 
 // UpdateState sets the internal state and triggers a pipeline update.
-func (m *mockStateProvider) UpdateState(state State, pipeline *PipelineImpl) {
+func (m *mockStateProvider) UpdateState(state optimistic_sync.State, pipeline *Pipeline) {
 	m.state = state
 	pipeline.OnParentStateUpdated(state)
 }
 
 // GetState returns the current internal state.
 
-func (m *mockStateProvider) GetState() State {
+func (m *mockStateProvider) GetState() optimistic_sync.State {
 	return m.state
 }
 
 // mockStateConsumer is a mock implementation used in tests to receive state updates from the pipeline.
 // It exposes a buffered channel to capture the state transitions.
 type mockStateConsumer struct {
-	updateChan chan State
+	updateChan chan optimistic_sync.State
 }
 
-var _ PipelineStateConsumer = (*mockStateConsumer)(nil)
+var _ optimistic_sync.PipelineStateConsumer = (*mockStateConsumer)(nil)
 
 // NewMockStateConsumer creates a new instance of mockStateConsumer with a buffered channel.
 func NewMockStateConsumer() *mockStateConsumer {
 	return &mockStateConsumer{
-		updateChan: make(chan State, 10),
+		updateChan: make(chan optimistic_sync.State, 10),
 	}
 }
 
-func (m *mockStateConsumer) OnStateUpdated(state State) {
+func (m *mockStateConsumer) OnStateUpdated(state optimistic_sync.State) {
 	m.updateChan <- state
 }
 
 // waitForStateUpdates waits for a sequence of state updates to occur or timeout after 500ms.
 // updates must be received in the correct order or the test will fail.
-func waitForStateUpdates(t *testing.T, updateChan <-chan State, errChan <-chan error, expectedStates ...State) {
+func waitForStateUpdates(t *testing.T, updateChan <-chan optimistic_sync.State, errChan <-chan error, expectedStates ...optimistic_sync.State) {
 	done := make(chan struct{})
 	unittest.RequireReturnsBefore(t, func() {
 		for _, expected := range expectedStates {
@@ -79,7 +80,7 @@ func waitForStateUpdates(t *testing.T, updateChan <-chan State, errChan <-chan e
 
 // waitNeverStateUpdate verifies that no state updates occur within 500ms.
 // The test fails if any unexpected state transition is observed.
-func waitNeverStateUpdate(t *testing.T, updateChan <-chan State, errChan <-chan error) {
+func waitNeverStateUpdate(t *testing.T, updateChan <-chan optimistic_sync.State, errChan <-chan error) {
 	done := make(chan struct{})
 	unittest.RequireNeverReturnBefore(t, func() {
 		select {
@@ -124,7 +125,7 @@ func waitForError(t *testing.T, errChan <-chan error, expectedErr error) {
 
 // createPipeline initializes and returns a pipeline instance with its mock dependencies.
 // It returns the pipeline, the mocked core, a state update channel, and the parent state provider.
-func createPipeline(t *testing.T) (*PipelineImpl, *osmock.Core, <-chan State, *mockStateProvider) {
+func createPipeline(t *testing.T) (*Pipeline, *osmock.Core, <-chan optimistic_sync.State, *mockStateProvider) {
 	mockCore := osmock.NewCore(t)
 	parent := NewMockStateProvider()
 	stateReceiver := NewMockStateConsumer()
